Move ReasonerTest to correct location such that `resolverRegistry()` can be package-private again.
