+++ test/integration/reasoner/BUILD
@@ -19,6 +19,31 @@ load("@graknlabs_dependencies//tool/checkstyle:rules.bzl", "checkstyle_test")
 load("@graknlabs_dependencies//builder/java:rules.bzl", "host_compatible_java_test")
 
 host_compatible_java_test(
+    name = "test-reasoner",
+    srcs = ["ReasonerTest.java"],
+    test_class = "grakn.core.reasoner.ReasonerTest",
+    native_libraries_deps = [
+        "//rocks:rocks",
+        "//logic:logic",
+        "//:grakn",
+        "//concept:concept",
+    ],
+    deps = [
+        # Internal dependencies
+        "//test/integration/util:util",
+        "//common:common",
+        "//concurrent:concurrent",
+
+        # External dependencies from Grakn Labs
+        "@graknlabs_graql//java:graql",
+    ],
+    resources = [
+        "//common/test:logback"
+    ],
+    resource_strip_prefix = "common/test",
+)
+
+host_compatible_java_test(
     name = "test-resolution",
     srcs = ["ResolutionTest.java"],
     native_libraries_deps = [
