+++ test/integration/reasoner/ReasonerTest.java
@@ -13,10 +13,9 @@
  *
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- *
  */
 
-package grakn.core.test.integration;
+package grakn.core.reasoner;
 
 import grakn.core.common.parameters.Options;
 import grakn.core.concurrent.actor.EventLoopGroup;
